# gh-136681: optimize builtin calls if kwargs allowed