--- a/Misc/NEWS.d/next/Core_and_Builtins/2025-07-16-14-44-09.gh-issue-136681.h1s2fH.rst
+++ b/Misc/NEWS.d/next/Core_and_Builtins/2025-07-16-14-44-09.gh-issue-136681.h1s2fH.rst
@@ -1,0 +1,2 @@
+Optimize calling builtins, accepting positional or keyword arguments in case
+no keyword arguments provided.  Patch by Sergey B Kirpichev.