… Lib/test/clinic.test.c
… Lib/test/test_clinic.py
… Modules/_io/clinic/textio.c.h
… Modules/_io/clinic/winconsoleio.c.h
… Modules/_sqlite/clinic/connection.c.h
… Modules/_sqlite/clinic/cursor.c.h
… Modules/_sqlite/clinic/module.c.h
… Modules/_sre/clinic/sre.c.h
… Modules/_ssl/clinic/cert.c.h
… Modules/_testcapi/clinic/exceptions.c.h
… Modules/_zstd/clinic/_zstdmodule.c.h
… Modules/_zstd/clinic/compressor.c.h
… Modules/_zstd/clinic/decompressor.c.h
… Modules/_zstd/clinic/zstddict.c.h
… Modules/cjkcodecs/clinic/multibytecodec.c.h
… Modules/clinic/_asynciomodule.c.h
… Modules/clinic/_bisectmodule.c.h
… Modules/clinic/_bz2module.c.h
… Modules/clinic/_codecsmodule.c.h
… Modules/clinic/_collectionsmodule.c.h
… Modules/clinic/_csv.c.h
… Modules/clinic/_cursesmodule.c.h
… Modules/clinic/_datetimemodule.c.h
… Modules/clinic/_elementtree.c.h
… Modules/clinic/_functoolsmodule.c.h
… Modules/clinic/_hashopenssl.c.h
… Modules/clinic/_lsprof.c.h
… Modules/clinic/_lzmamodule.c.h
… Modules/clinic/_opcode.c.h
… Modules/clinic/_pickle.c.h